# practices/practice_04/omnixon/omnixon-library/tests/server_openapi.json
{
  "components": {
    "schemas": {
...
      "Attachment": {
        "description": "A file for the model: give either `url` or `data` (base64).",
        "properties": {
          "data": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "the file, base64-encoded",
            "title": "Data"
          },
          "media_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "e.g. image/png or application/pdf; required with `data`, guessed from the URL otherwise",
            "title": "Media Type"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "a file name, shown in the history",
            "title": "Name"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "http(s) URL the model provider can fetch",
            "title": "Url"
          }
        },
        "title": "Attachment",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Interrupted": {
        "properties": {
          "interrupted": {
            "title": "Interrupted",
            "type": "boolean"
          },
          "text": {
            "default": "",
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "interrupted"
        ],
        "title": "Interrupted",
        "type": "object"
      },
...
      "MessageRequest": {
        "properties": {
          "attachments": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/Attachment"
            },
            "title": "Attachments",
            "type": "array"
          },
          "chat_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chat Id"
          },
          "request": {
            "title": "Request",
            "type": "string"
          },
          "save_message": {
            "default": true,
            "title": "Save Message",
            "type": "boolean"
          },
          "trace": {
            "default": false,
            "title": "Trace",
            "type": "boolean"
          },
          "use_memo": {
            "default": true,
            "title": "Use Memo",
            "type": "boolean"
          },
          "user_id": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          }
        },
        "required": [
          "request"
        ],
        "title": "MessageRequest",
        "type": "object"
      },
      "MessageResponse": {
        "properties": {
          "chat_id": {
            "title": "Chat Id",
            "type": "integer"
          },
          "response": {
            "title": "Response",
            "type": "string"
          },
          "trace": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/TraceStep"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "required": [
          "response",
          "user",
          "chat_id"
        ],
        "title": "MessageResponse",
        "type": "object"
      },
      "Model": {
        "properties": {
          "base_url": {
            "default": "https://openrouter.ai/api/v1",
            "title": "Base Url",
            "type": "string"
          },
          "has_api_token": {
            "description": "Whether the model has a key of its own (the key itself is never shown).",
            "readOnly": true,
            "title": "Has Api Token",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "default": "",
            "title": "Name",
            "type": "string"
          },
          "request_json": {
            "additionalProperties": true,
            "title": "Request Json",
            "type": "object"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          },
          "use_proxy": {
            "default": true,
            "title": "Use Proxy",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "request_json",
          "timestamp",
          "has_api_token"
        ],
        "title": "Model",
        "type": "object"
      },
      "ModelCreate": {
        "properties": {
          "api_token": {
            "anyOf": [
              {
                "maxLength": 512,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Token"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "request_json": {
            "additionalProperties": true,
            "title": "Request Json",
            "type": "object"
          },
          "use_proxy": {
            "default": true,
            "title": "Use Proxy",
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "request_json"
        ],
        "title": "ModelCreate",
        "type": "object"
      },
      "ModelUpdate": {
        "properties": {
          "api_token": {
            "anyOf": [
              {
                "maxLength": 512,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Token"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "request_json": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Json"
          },
          "use_proxy": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Use Proxy"
          }
        },
        "title": "ModelUpdate",
        "type": "object"
      },
...
      "TraceStep": {
        "properties": {
          "args": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Args"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "input_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Tokens"
          },
          "kind": {
            "enum": [
              "model",
              "tool"
            ],
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "output_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Tokens"
          },
          "result": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "step": {
            "title": "Step",
            "type": "integer"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          }
        },
        "required": [
          "step",
          "kind",
          "name"
        ],
        "title": "TraceStep",
        "type": "object"
      },
      "User": {
        "properties": {
          "agent_id": {
            "title": "Agent Id",
            "type": "integer"
          },
          "external_id": {
            "title": "External Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "id",
          "agent_id",
          "external_id",
          "timestamp"
        ],
        "title": "User",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/models": {
      "get": {
        "operationId": "get_models_api_v1_admin_models_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Model"
                  },
                  "title": "Response Get Models Api V1 Admin Models Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List all models",
        "tags": [
          "Admin API"
        ]
      },
      "post": {
        "operationId": "create_model_api_v1_admin_models_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Model"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create a new model",
        "tags": [
          "Admin API"
        ]
      }
    },
    "/api/v1/admin/models/{model_id}": {
      "delete": {
        "operationId": "delete_model_api_v1_admin_models__model_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "model_id",
            "required": true,
            "schema": {
              "title": "Model Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Model"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete a model",
        "tags": [
          "Admin API"
        ]
      },
      "get": {
        "operationId": "get_model_api_v1_admin_models__model_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "model_id",
            "required": true,
            "schema": {
              "title": "Model Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Model"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get a model by ID",
        "tags": [
          "Admin API"
        ]
      },
      "patch": {
        "operationId": "update_model_api_v1_admin_models__model_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "model_id",
            "required": true,
            "schema": {
              "title": "Model Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Model"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update a model",
        "tags": [
          "Admin API"
        ]
      }
    },
...
    "/api/v1/request": {
      "post": {
        "operationId": "request_api_v1_request_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Send a message and get an AI response",
        "tags": [
          "Main API"
        ]
      }
    },
    "/api/v1/request-stream": {
      "post": {
        "description": "Server-sent events: `event: user` with the user, then one `data:` event per\ntext chunk (a JSON string), then `event: done` with the user, or `event: error`\nwith `{\"detail\": ...}`. With `trace: true`, `event: trace` events (one step of the\nchain of calls each) come in between, as the steps finish. A stream that is stopped (by `POST /users/{id}/interrupt`, or because the same user\nsent a new request) ends with `event: interrupted` (the user) instead of `done`; what was said so far is in the history.",
        "operationId": "request_stream_api_v1_request_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Send a message and get an AI response via SSE",
        "tags": [
          "Main API"
        ]
      }
    },
...
    "/api/v1/users/{user_id}/interrupt": {
      "post": {
        "description": "Ends the stream of this user with the agent of the token: the stream gets `event: interrupted`, what was said so far is\nsaved to the history (marked `interrupted`) and to the memory, and then this returns. A new request of the same user does\nthe same by itself, so this is for stopping without saying anything new.",
        "operationId": "interrupt_user_api_v1_users__user_id__interrupt_post",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Interrupted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stop the answer that is being streamed to a user",
        "tags": [
          "Main API"
        ]
      }
    },
...
  }
}